Relocate the Kube API
Match the conventions of latest Kubebuilder.
