+++ .dockerignore
@@ -7,5 +7,3 @@
 !/go.mod
 !/go.sum
 !/main.go
-
-!/pkg
