+++ Dockerfile
@@ -11,9 +11,9 @@ RUN go mod download
 
 # Copy the go source
 COPY main.go main.go
+COPY api/ api/
 COPY controllers/ controllers/
 COPY internal/ internal/
-COPY pkg/ pkg/
 COPY redskyapi/ redskyapi/
 
 # Build
