+++ internal/server/server_test.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 	"time"
 
-	redskyv1alpha1 "github.com/redskyops/redskyops-controller/pkg/apis/redsky/v1alpha1"
+	redskyv1alpha1 "github.com/redskyops/redskyops-controller/api/v1alpha1"
 	"github.com/redskyops/redskyops-controller/redskyapi/experiments/v1alpha1"
 	redskyapi "github.com/redskyops/redskyops-controller/redskyapi/experiments/v1alpha1"
 	"github.com/stretchr/testify/assert"
