+++ internal/trial/metadata.go
@@ -19,7 +19,7 @@ package trial
 import (
 	"strings"
 
-	redskyv1alpha1 "github.com/redskyops/redskyops-controller/pkg/apis/redsky/v1alpha1"
+	redskyv1alpha1 "github.com/redskyops/redskyops-controller/api/v1alpha1"
 )
 
 // GetInitializers returns the initializers for the specified trial
