+++ internal/trial/status_test.go
@@ -18,7 +18,7 @@ package trial
 import (
 	"testing"
 
-	redskyv1alpha1 "github.com/redskyops/redskyops-controller/pkg/apis/redsky/v1alpha1"
+	redskyv1alpha1 "github.com/redskyops/redskyops-controller/api/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 )
