+++ redskyctl/internal/commands/check/experiment.go
@@ -21,8 +21,8 @@ import (
 	"io/ioutil"
 	"strings"
 
+	redskyv1alpha1 "github.com/redskyops/redskyops-controller/api/v1alpha1"
 	"github.com/redskyops/redskyops-controller/internal/template"
-	redskyv1alpha1 "github.com/redskyops/redskyops-controller/pkg/apis/redsky/v1alpha1"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commander"
 	"github.com/spf13/cobra"
 	batchv1 "k8s.io/api/batch/v1"
